Makes demo usable
This fixes quite a few things so that you can actually see things work!
