+++ README.md
@@ -1,10 +1,22 @@
-WIP
+# Basic example starting a trace in the browser and propagating it to a node.js server
 
-Startup zipkin on localhost, then start the node server
+First, run the frontend, which will listen on http://localhost:8081
+
 ```bash
-$ browserify browser.js -o public/bundle.js 
-$ node server.js 
+$ node frontend.js
 ```
 
-Hit http://localhost:3000/ a couple times (note there's no content)
-Look at zipkin http://localhost:9411/?serviceName=browser
+Next, run [Zipkin](http://zipkin.io/), which stores and queries traces reported by the above services.
+
+```bash
+$ wget -O zipkin.jar 'https://search.maven.org/remote_content?g=io.zipkin.java&a=zipkin-server&v=LATEST&c=exec'
+$ java -jar zipkin.jar
+```
+
+Now, build javascript for the browser
+
+```bash
+$ browserify browser.js -o bundle.js
+```
+
+Finally, `open index.html` and reload a couple times. You can see traces in zipkin here: http://localhost:9411/?serviceName=frontend
