+++ index.html
@@ -1,0 +1,5 @@
+<html>
+  <body>
+    <script type="text/javascript" charset="utf-8" src="./bundle.js"></script>
+  </body>
+</html>
