+++ package.json
@@ -8,11 +8,10 @@
   },
   "dependencies": {
     "express": "^4.14.0",
-    "express-http-proxy": "^0.9.1",
-    "zipkin": "^0.2.6",
-    "zipkin-instrumentation-express": "^0.2.6",
-    "zipkin-instrumentation-fetch": "^0.2.6",
-    "zipkin-transport-http": "^0.2.6"
+    "zipkin": "^0.2.7",
+    "zipkin-instrumentation-express": "^0.2.7",
+    "zipkin-instrumentation-fetch": "^0.2.7",
+    "zipkin-transport-http": "^0.2.7"
   },
   "devDependencies": {
     "eslint": "^3.4.0",
